HUD (Working Inventory System)
The HUD widget was created and an inventory was added (Current Ball Variation Count = 3). The blueprint holds the necessary events to change ball count and select/deselect a ball as well as animated states for inventory UI elements. No Integrations were made between the Player Controller and HUD widget (Widget stands alone), but a throwaway project was used to verify code functionality and future integrations.