Main Menu & Options Menu added
Add the main menu widget and the options menu widget. The exit game button is fully functional. The play button opens the current WIP level. The options button takes you to the empty Options menu.
Also added is a menu player controller that always toggles the mouse cursor, and a main menu game mode, so it does not interfere with the main game mode of the project.